feat: load modules based on remoteConfig
Remote Config is a seperate product within the firebase platform
which allows users to store state remotely which can be edited
outside of deployments.
Link: https://firebase.google.com/docs/remote-config

Using remote configuration we are then able to defer
module configuration out to the target environment.

This commit adds a check against remote config when:
1. Building the desktop menu
2. Registering routes
3. Example of checking module support within a Profile page

Questions which still require further investigation:
* Loading state when fetching remote configuration
* Use mobx as a wrapper for this Firebase product

Request for comment:
* A new src/modules/ directory to serve as the primary
entry point for interacting with the Module concept.